% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acc_csv.R
\name{idle_na_locf}
\alias{idle_na_locf}
\title{Replace Values with Last Observation Carried Forward from Idle Sleep Mode}
\usage{
idle_na_locf(df)
}
\arguments{
\item{df}{An object with columns \code{X}, \code{Y}, and `Z``}
}
\value{
A data set with the zeros filled in
}
\description{
Replace Values with Last Observation Carried Forward from Idle Sleep Mode
}
\examples{
df = data.frame(
  X = c(0.3/sqrt(0.5), rep(0, 3)),
  Y = c(0.4/sqrt(0.5), rep(0, 3)),
  Z = c(0.5/sqrt(0.5), rep(0, 3)),
  stringsAsFactors = FALSE)
idle_na_locf(df)
}
