% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_read_gt3x.R
\name{impute_zeros}
\alias{impute_zeros}
\title{Impute Zeros and Missing data}
\usage{
impute_zeros(data, dates, header)
}
\arguments{
\item{data}{data output from \code{\link{py_read_gt3x}}}

\item{dates}{dates vector from \code{\link{py_read_gt3x}}}

\item{header}{header metadata from \code{\link{py_read_gt3x}}}
}
\value{
A tibble of the data with zeros
}
\description{
Impute Zeros and Missing data
}
