% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pygt3x_required_modules.R
\docType{data}
\name{pygt3x_required_modules}
\alias{pygt3x_required_modules}
\title{Python Modules Required for \code{pygt3x} Package}
\format{
A character vector of Python modules required to run the scripts
}
\usage{
pygt3x_required_modules
}
\description{
Python Modules Required for \code{pygt3x} Package
}
\keyword{datasets}
