% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_read_gt3x.R
\name{py_read_gt3x}
\alias{py_read_gt3x}
\title{Read GT3X File}
\usage{
py_read_gt3x(file, create_time = FALSE)
}
\arguments{
\item{file}{file to GT3X file}

\item{create_time}{Should time stamps be created?}
}
\value{
A list of meta data, accelerometry data, and time stamps
}
\description{
Read GT3X File
}
